# Optimize Bytes::slice for short slices